requirements.txt
PyGithub
flask-restful
numpy
packaging
pillow